Adjust timer settings to be more accurate
T3_CK_SE should be 12bit register, and the upper 4bits are ignored
and assumed as all-zeros. So, the counter ran at around 16kHz
in spite of the comment says it does at 1kHz.

Other code expects it runs at 16kHz for some reasons I could not remember,
but maybe it was based on observed behaviors. So, it just works
regardless of this fix, but now it runs more precisely.

The interrupt handler expects per-sec interrupt, and the END value seems
to have adjusted also based on the observed behavior. Now this END value
is also adjusted to work with the precised T3_CK_SE value.

timer3_example is added to check behaviors easily.